requirements.txt
streamlit>=1.30.0
pyswisseph>=2.10.0
pytz>=2024.1
